remove playground and index pages, add redirects, update navigation
