+++ gatsby-node/create-redirects.js
@@ -1,8 +1,16 @@
 // This param is used when `/[tutorials, how-to-guides]/` are then redirected to
 // `/[tutorials, how-to-guides]/<defaultVersion>/`.
-const defaultVersion = 'v3';
+//const defaultVersion = 'v3';
 
 const redirects = [
+  {
+    fromPath: '/',
+    toPath: `/quick-start/`,
+  },
+  {
+    fromPath: '/playground/',
+    toPath: `https://substrate.io/developers/playground/`,
+  },
   {
     fromPath: '/how-to-guides/',
     toPath: `/reference/`,
